my-first-app/requirements.txt
psycopg2-binary
flask
Flask-SQLAlchemy
flask-migrate